feat: Fix dates in show my results view

+++ app/src/main/java/com/panco/multichoice/models/Game.kt
@@ -1,3 +1,3 @@
 package com.panco.multichoice.models
 
-data class Game(var gameId: Int, var dateStarted: Int, var playerId: Int, var score: Int)
+data class Game(var gameId: Int, var dateStarted: Long, var playerId: Int, var score: Int)
